+++ tests/Jroc.Test262.Tests/language/expressions/class/elements/ExecutionTests.cs
@@ -6,6 +6,18 @@ public class ExecutionTests : ExecutionTestsBase
 {
     public ExecutionTests() : base("language.expressions.class_.elements") { }
 
+    [Fact(DisplayName = "after-same-line-static-gen-private-method-usage")]
+    public Task after_same_line_static_gen_private_method_usage()
+        => ExecutionTest("after-same-line-static-gen-private-method-usage");
+
+    [Fact(DisplayName = "after-same-line-static-method-private-method-usage")]
+    public Task after_same_line_static_method_private_method_usage()
+        => ExecutionTest("after-same-line-static-method-private-method-usage");
+
+    [Fact(DisplayName = "after-same-line-static-method-private-names")]
+    public Task after_same_line_static_method_private_names()
+        => ExecutionTest("after-same-line-static-method-private-names");
+
     [Fact(DisplayName = "class-name-static-initializer-anonymous")]
     public Task class_name_static_initializer_anonymous()
         => ExecutionTest("class-name-static-initializer-anonymous");
@@ -37,6 +49,14 @@ public class ExecutionTests : ExecutionTestsBase
     [Fact(DisplayName = "fields-multiple-definitions-static-private-methods-proxy")]
     public Task fields_multiple_definitions_static_private_methods_proxy()
         => ExecutionTest("fields-multiple-definitions-static-private-methods-proxy");
+
+    [Fact(DisplayName = "multiple-definitions-private-method-usage")]
+    public Task multiple_definitions_private_method_usage()
+        => ExecutionTest("multiple-definitions-private-method-usage");
+
+    [Fact(DisplayName = "multiple-definitions-private-names")]
+    public Task multiple_definitions_private_names()
+        => ExecutionTest("multiple-definitions-private-names");
 
     [Fact(DisplayName = "new-sc-line-method-computed-symbol-names")]
     public Task new_sc_line_method_computed_symbol_names()